Feature/admin dealership management

+++ src/app/modules/vehicles-listing/services/vehicle-fax-report.service.ts
@@ -41,7 +41,7 @@ export class VehicleFaxReportService implements ServiceInterface {
 
   async store(
     req: Request,
-    dto: { id: number; file: any },
+    dto: { id: number; file: Express.Multer.File },
   ): Promise<Record<string, any>> {
     const queryRunner =
       this.vehicleFaxReportRepository.manager.connection.createQueryRunner();
@@ -86,15 +86,16 @@ export class VehicleFaxReportService implements ServiceInterface {
       const folder = `vehicle/fax/${vechicle_id}`;
 
       // Upload attachment file
-      const fileName = dto.file.originalname;
-      const fileStream = Readable.from(dto.file.buffer);
-      const fileSize = dto.file.size;
+      const file = dto.file;
+      const fileName = `${Date.now()}-${file.originalname}`;
+      const fileStream = Readable.from(file.buffer);
+      const key = `${folder}/${fileName}`;
 
-      const newFile = await this.fileUploadService.uploadFileStream(
+      const newFile = await this.fileUploadService.uploadStream(
+        key,
         fileStream,
-        fileName,
-        fileSize,
-        folder,
+        file.mimetype,
+        file.size,
       );
 
       uploadedFiles = `${folder}/${newFile}`;
@@ -111,7 +112,9 @@ export class VehicleFaxReportService implements ServiceInterface {
         // Delete old vehicle fax report and update new one and merge
         // -------------
         // Delete old Fax report
-        await this.fileUploadService.deleteFile(vehicleFaxReport.attachment);
+        if (vehicleFaxReport.attachment) {
+          await this.fileUploadService.deleteFile(vehicleFaxReport.attachment);
+        }
 
         // Merge vehicle Fax report with new fax file attachment
         vehicleFaxReport = queryRunner.manager.merge(
@@ -155,6 +158,8 @@ export class VehicleFaxReportService implements ServiceInterface {
       }
       this.logger.error(error);
       return throwCatchError(error);
+    } finally {
+      await queryRunner.release();
     }
   }
 
@@ -218,7 +223,9 @@ export class VehicleFaxReportService implements ServiceInterface {
       }
 
       // try delete the attachment from s3
-      await this.fileUploadService.deleteFile(vehicleFaxReport.attachment);
+      if (vehicleFaxReport.attachment) {
+        await this.fileUploadService.deleteFile(vehicleFaxReport.attachment);
+      }
 
       // delete attachment from database Record
       await queryRunner.manager.delete(VehicleFaxReport, id);
@@ -230,8 +237,11 @@ export class VehicleFaxReportService implements ServiceInterface {
         message: `Attachment removed successfully`,
       };
     } catch (error) {
+      await queryRunner.rollbackTransaction();
       this.logger.error(error);
       return throwCatchError(error);
+    } finally {
+      await queryRunner.release();
     }
   }
 
@@ -267,12 +277,24 @@ export class VehicleFaxReportService implements ServiceInterface {
         vehicleFaxReport = queryRunner.manager.create(VehicleFaxReport, {
           vehicle_id: id,
           expired_at: new Date(Date.now() + 60 * 24 * 60 * 60 * 1000), // add 60 days
+          status: VehicleFaxReportStatus.REQUESTED,
         });
       } else {
+        // Check if user already requested for carfax report
+        if (
+          vehicleFaxReport.status === VehicleFaxReportStatus.REQUESTED ||
+          vehicleFaxReport.status === VehicleFaxReportStatus.PENDING
+        ) {
+          return {
+            message: 'You have already requested for carfax report',
+          };
+        }
         // Delete old vehicle fax report and update new one and merge
         // -------------
         // Delete old Fax report
-        await this.fileUploadService.deleteFile(vehicleFaxReport.attachment);
+        if (vehicleFaxReport.attachment) {
+          await this.fileUploadService.deleteFile(vehicleFaxReport.attachment);
+        }
 
         // Merge vehicle Fax report with new fax file attachment
         vehicleFaxReport = queryRunner.manager.merge(
@@ -280,9 +302,13 @@ export class VehicleFaxReportService implements ServiceInterface {
           vehicleFaxReport,
           {
             expired_at: new Date(Date.now() + 60 * 24 * 60 * 60 * 1000), // add 60 days
+            status: VehicleFaxReportStatus.REQUESTED,
+            attachment: null,
           },
         );
       }
+
+      await queryRunner.manager.save(VehicleFaxReport, vehicleFaxReport);
 
       // Add extraction task to queue
       await this.vehicleQueue.add('vehicle-fax-report-apply', vehicleFaxReport);
@@ -291,11 +317,14 @@ export class VehicleFaxReportService implements ServiceInterface {
       await queryRunner.commitTransaction();
 
       return {
-        message: `Attachment removed successfully`,
+        message: `CarFax report applied successfully`,
       };
     } catch (error) {
+      await queryRunner.rollbackTransaction();
       this.logger.error(error);
       return throwCatchError(error);
+    } finally {
+      await queryRunner.release();
     }
   }
